backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["magnus <magnus@dahleide.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
SQLAlchemy = "^2.0.27"
uvicorn = "^0.27.1"
python-jose = "^3.3.0"
passlib = {extras = ["brypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
